backend/pyproject.toml
[tool.poetry]
name = "koala"
version = "0.1.0"
description = ""
authors = ["Michel Meyer <meyer_michel@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
dataclasses-json = "^0.5.7"
dependency-injector = "^4.40.0"
immudb-py = "^1.3.2"
fastapi = "^0.87.0"
uvicorn = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-cov = "3.0.0"
mypy = "0.961"
flake8 = "4.0.1"
black = "22.6.0"
isort = "5.10.1"
bandit = "1.7.4"
pylint = "2.14.4"
pytest-md-report = "^0.3.0"

[tool.mypy]
files = "koala/**/*.py"
allow_untyped_decorators = false
warn_unused_configs = true
allow_subclassing_any = false
allow_untyped_calls = false
allow_untyped_defs = false
allow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = false
pretty = true
exclude = "tests/*"

[[tool.mypy.overrides]]
module = 'immudb.*'
ignore_missing_imports = true


[tool.pylint.format]
max-line-length = 200
disable = ["too-few-public-methods"]
extension-pkg-whitelist= ["dependency_injector.containers", "dependency_injector.providers"]

[tool.pylint.messages_control]
disable = ''',
    fixme,
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring,
 '''

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
koala= "koala.main:main"


